refactored HUD debug panel into its own script
