+++ time-arena-game/Assets/Assets/Materials/Crystals/Crystal_Overlay.mat
@@ -92,7 +92,7 @@ Material:
     - Vector1_c49ac3afe4db4b9282ca8481a4695e49: 5
     - Vector1_dbe12339d960441e9fe119160eec01cd: 1
     - Wave_Freq: 5
-    - Wave_Incr: 15.07734
+    - Wave_Incr: 71.69204
     - Wave_Width: 1
     - _AlphaClip: 0
     - _Blend: 0
