+++ time-arena-game/Assets/Assets/Resources/Player.prefab
@@ -9196,6 +9196,7 @@ GameObject:
   - component: {fileID: 5590288987897476469}
   - component: {fileID: 71524245628701223}
   - component: {fileID: 4621444560055183819}
+  - component: {fileID: 2539957824911093140}
   m_Layer: 5
   m_Name: UI
   m_TagString: Untagged
@@ -9401,6 +9402,20 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _container: {fileID: 8560333854563395425}
   _text: {fileID: 8006470102416607322}
+--- !u!114 &2539957824911093140
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6146106990117875522}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3bfd416579202470d8e527284f03682e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _debugCanvasGroup: {fileID: 1999155244747893999}
+  _text: {fileID: 4281627418019176230}
 --- !u!1 &6189879322987748197
 GameObject:
   m_ObjectHideFlags: 0
@@ -10009,6 +10024,7 @@ MonoBehaviour:
   _cooldowns: {fileID: 5590288987897476469}
   _winningDisplay: {fileID: 71524245628701223}
   _score: {fileID: 4621444560055183819}
+  _debugPanel: {fileID: 2539957824911093140}
   View: {fileID: 5690578699466181193}
   TeamDispl: {fileID: 8151362755639177340}
   DebugCanvasGroup: {fileID: 1999155244747893999}
