+++ time-arena-game/Assets/Scripts/Hud/HudDebugPanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3bfd416579202470d8e527284f03682e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
